Lazy partial rendering for agenda view
The agenda can be very very big and the amount of data on a page can
be too much for humans and for computers to efficiently process.  This
lazy loading should help in such cases where the majority of the data
is only fetched when the item is on screen.