Moving from PODs to current modern EmberJS structure
This marks a move from Ember components embedded in PODs to Ember
components embedded in the regular folder structure.  The benefit of
this is that the file name is not "template.hbs" and "component.js" but
rather a sensible name like "agenda-header.hbs" and "agenda-header.js".

The structure also matches modern Ember guidelines so it will help with
upgrades in the distant future.