Correct data-test invocation in curly components
Manual of ember-test-selectors indicates how to use these test
selectors.  I don't grok how this would have worked in the past, but
this puts the invocation in line with the manual of the previous and
current version of ember-test-selectors for curly components.

We should pass this information differently for the angle bracket
components too, but because we don't use testing frameworks within
EmberJS the current solution might actually work.  We could then fix in
a separate PR.