Allow to save changes on a subcase for a closed agenda
Behaviour is currently not the same depending changes to the subcase
are made via the agendaitem or via the subcase page directly.

If changes on the subcase should not be allowed after an agenda handling
the subcase has been closed, the edit functionality should not be made
available to the users in the first place.