Move agenda documents loading to controller to provide progress info
By moving the loading of agenda documents to the controller, we can
provide more detailed progress information to the user. Still
ensuring though that the agenda-overview component only starts
rendering when all data has been loaded.

Also loading documents in seperate requests per agendaitems. With a
semi-hot cache (which seems to be the case a lot in practice)
this yields a better performance than including all documents in
1 bulk agenda request.