Share grouped notas calculation between agendaitems index and detail
By restructuring the templates/controller/routes of
agenda.agendaitems.index and agenda.agendaitems.agendaitem routes the
data and calculation of grouped notas can be shared, resulting in less
loading states when navigating between subroutes of the same agenda.

Also making sure documents are only loaded when navigating to the
index route and not to the detail route of an agendaitem.