Reduce number of included resources in publications index model hook
Adding included resources reduces the number of requests sent to the
backend but also has a performance penalty: the executed query takes
longer and cache entry gets invalidated quickly due to the agressive
cache clearing.

Currently only including resources that are rather 'stable' to lower
chances of cache invalidation.

Loading additional data in separate requests also has a performance
benefit for user that display only a subset of the columns in the
publications data table. If a column is not shown the additional data
will not be loaded, while in the current implementation data is always loaded.