Remove update of file received date on change of translation-activity received date
Since the file received dates are not shown to the user, we're not
sure if it's the intention of the user to update the received date of
each file. Files may be uploaded in multiple batches, having different
received dates. This information gets lost if we blindly overwrite the
dates if the received date of the translation-activity gets updated.