Allow immediate setting of filter by pressing enter
This is potentially useful for the user, but also especially for the
Cypress tests so we don't need to wait for the debounce.

Also tell Cypress to explicitly wait for the loading state of the table
to be gone. Because now use an inline loading state as opposed to
ember's route based loading system, Cypress was trying to click on the
table rows too quickly, before the actual results had rendered, which
caused the tests to fail.