Tweak default sorting of agendas table & QP error
The session-kind test expects the PVV Ministerraad to come after the
normal Ministerraad in a filtered agendas table. With this new
default filtering, we explicitly apply this ordering.

An agendaitem delete call was causing errors which might be related to
these Ember bugs:

- https://github.com/emberjs/ember.js/issues/20051
- https://github.com/emberjs/ember.js/pull/19971

Leaving out setting anchor to null in hope that it might fix it.