Prevent multiple backend calls per subcase item
Previosuly, the subcase-item component was using `computed` properties
of the subcase model to side-load data (e.g. the phases belonging to a
subcase as well as the approval state of the subcase). This resulted in
multiple backend calls being made multiple times (presumably because
they got fired off again when the component re-rendered). To prevent
this, and keep the number of backend calls manageable, we side-load the
data using tasks which don't refire.

Where possible, the number of backend calls has also been reduced by
removing superfluous calls.