Refactor phases computed property on subcase
Fetching `agenda-activity` here seems unnecessary, because the backend
microservice will check itself if such activities exist. So in a case
where there are no agenda-activities, we will still only make 1 request
to the backend, and when there are activities, we will make 1 request
instead of 2.

Also adds a comment to the subcase phases post processing function.