Cypress subcase test, wait before deleting
Not the ideal solution, but it seems to fix an error that's caused by
Cypress immediately deleting the subcase while some requests are still
in flight.

Error at hand:

1) Subcase tests
    should add a subcase and then delete it:
   Error: The following error originated from your application code, not from Cypress.

> Attempted to handle event `pushedData` on <subcase:629675108AADBDDBE93A3D46> while in state root.deleted.inFlight.

I am guessing that the timing of the requests are off, and that perhaps
the way computed properties work make it so deleting the subcase record
while the requests are still in flight causes an error here.