Remove include param from publication overview model hook
Because visible columns and sorting are configurable by the enduser, a
lot of possible combinations are available. By making include depend
on the selected columns, chances that the cache will be hit are low.

By removing the include param from the request to the backend we
increase the chance to hit the cache to get the list of
publications. A lot more requests will be done to sideload related
data, but those (simple) requests will almost always hit the cache.