feat: split updatePreviousAccessLevels in two to handle one and all access levels
updatePreviousAccessLevels still updates all previous access levels as
before, new function updatePreviousAccessLevel only updates a single
access level. This is used when uploading a new version of a piece
(because we only want to affect the last version, not all of them).

updatePreviousAccessLevels behaviour is maintained, where it stops if it
encounters an Intern Secretarie or Ministerraad access level.