refactor: display governmentAreas in publication only if via MR
The previous code was, show governmentAreas always, but only make them
editable if the case is not via MR. Now, a case that is not via MR does
not have a decisionmakingFlow, so it will never have governmentAreas.
Thus, we now display governmentAreas only for cases via MR, and
read-only, because they're not meant to be editable in this view.