fix: apply overflow-wrap: anywhere to RDFa editor
When a line is too long and a user is typing in the middle of the
sentence, the editor doesn't always wrap long words at the end of the
line. To remedy this we apply overflow-wrap: anywhere to the editor,
which makes it break up words over multiple lines if they're too long.

This way, instead of growing until it reaches the max-width, when a line
is too long due to the