feat: add isDraftAccessLevel helper
This helper gets most of the needed data passed in as arguments instead
of side-loading them and causing extra requests every time we need to
determine the state of an access level. This does mean that we will need
to drill down props through a number of components (optionally using a
hash instead of bare properties, though the essence is the same).

The publication activities as well as agenda status & possible previous
agenda are still side loaded. These can also be passed in as arguments,
though this would result in four extra arguments, making the templates
even bulkier.