fix: load the selected orgs/roles via task in afterModel hook
Loading them in the task looked like it worked but it actually didn't.
Seems that the component gets constructed before the setupController
hook from the route is called. That means that the tasks loading the
records would run using empty arrays (i.e., as if no orgs/roles were
seleceted).