chore: reorder options of certain concepts calls so they match warmup
This makes it easier at a glance to know if the warmup service will hit
the calls or not. In particular the filter/sort/include/etc. options are
ordered alphabetically. For example sort will be at the end of the
queryParams, so we set it at the end as well to make it more obvious.