chore: hide extension of source and derived file unless explicit
Showing .docx when the derived is a .pdf and we open the PDF gets a
littel confusing. We just always show the piece name, and when it's
explicit what file we're working with (e.g. in the details panel, where
you replace the source file only), we show the appropriate extension.