chore: show extension again
If there's a derived file, show the derived extension, otherwise show
the source extension. I suspect this will break the tests because at
first render we show the source extension (because the derived file is
still in-flight).
But if it's a non-issue I will leave it as is, because checking for
the loading state of a relation is kind of annoying in Ember.