fix: pass correct piece length when creating new version
When generating a piece's version name, we use the number of pieces to
determine what suffix it gets (BIS, TER, ...). Before, we passed in the
number of pieces + 1, representing the newly uploaded version. Now,
because we do link the relation between document container and
piece (before, this had inverse: null), the pieces' array length
automatically updates as soon as we add the new version. This means that
we no longer need to add one to the number of pieces when generating the
name.