fix: don't expect documentContainer to exist when deleting a piece
Because the relation between piece and document container is now
explicit in the frontend, when we delete a piece that is being
created (e.g. in an upload modal, by deleting the piece that was just
uploaded) we shouldn't expect the documentContainer to exist. This is
because the piece might be removed from the documentContainer.pieces
array, which immediately breaks the piece.documentContainer link too.