fix: deal with null argument in DecisionResultPill
Using .get('property') works only if we pass in a proxy object (e.g.
from a relationship). This doesn't work when coming from the search
interface, there we pass in either a record or null if no decision
result exists. Explicitly checking for the nullability of the
decisionResultCode remediates this and makes it compatible.