fix: ensure the hierarchical back tab is always active
Previously the hierarchical back buttons always were a LinkTo component
with a route parameter that linked to a parent route. Because of this,
they were always considered active (because a parent route is always an
active route). Some tests relied on this button being active, so we
ensure it's always active even when not passing a route (and making the
button behave like a history back button).