feat: expand KDataTable to support other things than tables
KDataTable is now split into a KData component that handles pagination
and exposes a Table or a Cards component that the consumer can use to
display the data. The consumer can also just use the data itself row by
row and choose how to display it by using a contextual body (i.e.
<:body></:body>).