fix: work with the value of archivedOptions in case search route
The radio list passes the option's value to the controller when a user
selects the radio button (between the choices of hiding deleted cases,
showing deleted cases, or only showing deleted cases). The query param
also relies on it being a plain string (it would look weird if we passed
in a JSON object, although it's possible).

So we should ensure that the radio list uses this.archived as a value
instead of an object when checking the checked state.

This fixes a small issue where selecting an option from the radio list,
leaving the route, and re-entering the route leaves the radio list
without a selection.