ci: wait 10 minutes for Elastic indexation
The failing test fails on a search request, where data is expected but
it doesn't exist. I suspect the Elastic indexation is taking too long,
the first tests in the run are meant to create some data but it never
gets indexed in Elastic because the early tests are ongoing at the same
time as the indexation. But then by the time the failing test comes
around we can't find the necessary data.