feat: make minister-filter use interface with IDs instead of records
Also stops using the construction with a tracked _selected. Instead we
always use the passed in argument for checking the checked status of
checkbox and such. When we go to toggle a minister, we just make a local
copy of the list, toggle it there, then pass the copy to the consumer.
This way we still respect Ember's DDAU by not mutating the argument,
while stepping away from having internal state that did not react to
argument changes (once _selected was set, if the arg selected changed by
e.g. resetting the filters, updates to the arg would not make the filter
visibly update).