fix: sort agendaitems on type.position,number
Before we sorted on type,number, which does not do what we expect, it
sorts on the type relationship, which currently mu-cl-resources
interprets as "sort on each property of this relationship". Additionally
this results in faulty responses from the new construct-based fetches
branch. What we really want is to sort on the type's position, so let's
just do that instead.