feat: stop passing around markForSignature
We no longer need it, as there's no longer a need for solely marking a
piece. From now on, we always start the siging flow, so it makes more
sense to start that process in the specific component that is
responsible for it. Passing around these actions no longer makes sense.