feat: adapt Mandatees::CheckboxList to support partial matches
The component normally only marks a minister as checked if all their
mandatees are selected. In the case of selecting signers we don't want
that, because we don't want to select the PM with both PM and regular
minister mandatees. The extra option tells the component to show a
minister as checked even if only some of the mandatees are selected.