feat: don't serialize outgoing secretary relations from user model
Although these relationships are hasMany on this side and don't get
serialized by Ember Data by default, we explicitly disable them here.
This is to prevent possible issues where someone editing a user in the
Gebruikersbeheer tab might accidentally delete the relationships. We
only ever want to set this relationship on the agenda/decision-activity
side.