feat: add button to remove sign flows in op te starten lijst
We wrap the whole template inside an if because the delete operation is
deleting the records we're showing, and ember throws errors about this.
E.g. as we delete the link between the sign flow and piece, the piece
we're rendering in the list is "gone" and we try to render a link that
doesn't exist. This throws an error. Other such errors also occurs. The
"simplest" fix is to simply show a loading state while this happens so
that we're not trying to render faulty data.