chore: bump ember-data to minor v4.8.0
It's good that we match the minor of Ember we're using elsewhere, but we
also specifically need it to resolve a deprecation, which was fixed in
this PR: https://github.com/emberjs/data/pull/8435

TL;DR: accessing promise proxies, which are returned by ED in stuff like
store.query and such is deprecated, as they will soon become regular
promises. Ember concurrency was accessing these proxies in some way, but
this was triggering the deprecation.
See: https://github.com/machty/ember-concurrency/issues/463