feat: show piece edit dropdown when a marked sign flow exists
We still hide it for any sign flows that's further along (i.e. on SH).
We also ensure that the dropdown only renders after we've loaded the
signature data. This to prevent some content flashing, as before the
dropdown would show for a split second while the sign flow data was
being loaded.