feat: extract updatePreviousAccessLevel from document-card
At least in the agendaitem view, we are refreshing the page after adding
a piece. Although up until now the call to update the access levels
that's in the document card seems to have gone through (even though the
component where the call originates from was re-rendered!) it seems
safer to extract this call. We're leaving in the document-card just in
case though.