feat: only show dropdown for marked pieces with a decision activity
In practice this means that the piece will only be editable in the
agendaitem documents route. This is because from e.g. the subcase view
we don't directly have the decision activity, which means that something
like uploading a new version (which means we need to create a new sign
flow) will fail. So we just prevent the behaviour instead.