fix: don't reload whole agendaitems route when searching
Full reload (and showing the loading state) means that we de-render the
search field and that users can't automatically keep typing.

This commit should still preserve the fix intended in KAS-4129, though
without making use of router.transitionTo and just using the
refreshModel behaviour of the queryParams again. Setting filter as a
tracked field instead of using `set()` seems to fix the behaviour.