fix: approvers + signature pill shouldn't be clickable in ongoing list
Approver's address was being passed incorrectly to the createRecord
call, as the property is called `approver` but (implicitly)
`approverLowerCase` was being passed. This caused the sign flow to have
no approver in SH.

The built-in procedure to check if a signature pill is clickable is fine
but we still need to be able to override it (and disable it) in certain
cases, specifically in the ongoing signatures list.