chore: don't make a count call in queryAll, get count from first call
Store#count is really just requesting a single record from
mu-cl-resources and using the returned metadata to know how many records
there are in total. While this is fine if you only care about the count,
since we're going to start fetching the actual records in
Store#queryAll, we're better off doing an initial call for the first
batch, checking the total count and then deciding if we need to fetch
further batches.

While this isn't a giant amount of performance gain, for e.g. the
government fields, the uncached count call takes ~200ms, which isn't
nothing.