chore: redirect bare agenda route to agenda.agendaitems
The bare agenda route is empty. Although nothing in our app links to it,
users might arrive there by (accidentally) removing /agendapunten or
/documenten from the agenda route. Or users might expect a working view
when manually navigating to /vergadering/:id/agenda/:id, which we
currently don't offer.

In any case instead of having a barren route that users might stumble
upon, it's better to redirect to a route we actually have populated.