fix: use correct intl string when report generation fails
We don't store a reason/message when a job fails so we have nothing to
show to users, so we shouldn't try to generate a message that takes in a
reason argument, otherwise the intl library generates an error.

We *should* store reasons at some point, but we don't, this is just a
patch to not have exceptions bubbling.