chore: move SubcasesOverviewHeader to child routes
The add-subcase route is also a child of subcases, if we put the header
in the parent we will also see it in the route to add a subcase and we
don't want it there. Instead we move it to the child routes where we
actually need it, i.e. the index route, which is just there to redirect
or show the empty state, and the subcase route, which has the data we
want to show.

The side nav is also moved to the subcase route as that is the only
place we actually use it.