fix: calculate agenda item number before saving type change
If we do it afterwards, we will potentially get the wrong number because
the query is counting this subcase too. E.g. if this subcase is added at
the end of the list after changing the type (because there are more
nota's than mededelingen) then we will get the subcase's number itself,
instead of the next number for mededelingen.