fix: use generated piece names when sorting
In the Agenda Check modal we sort the pieces of an agendaitem.
For regular agendaitems this works well as they all have positions
and the sorting can be done using the positions. For agendaitems
where we show a ratification this is not the case as the ratification
does not have a position.

To prevent a wrong sorting of the pieces in this modal, we will use the
generated names for sorting the pieces, ensuring that the sorting
algorithm uses the generated VR names.